requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9          # form data support

# ── Database ─────────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0                # async SQLite driver (swap for asyncpg in prod)
# asyncpg==0.29.0                # uncomment for PostgreSQL

# ── Auth & Security ───────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4            # password hashing
cryptography==42.0.8

# ── Validation ────────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0
email-validator==2.1.1

# ── Rate Limiting ─────────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Security Headers / CSRF ───────────────────────────────────────────────────────
secure==0.3.0                     # HTTP security headers

# ── Logging & Observability ───────────────────────────────────────────────────────
structlog==24.2.0

# ── Testing ───────────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0                     # async test client
